requirements.txt
requests
twitter
python-dotenv